interview auto-save (during summarization)
